Change the project name.
Aiming not to have the company name in it (because it is not related at
all with that company), I have changed the project name from "CSG
Library Management" to LibraryManagerApi (same name as the folder).

Adds the "aws-windows-deployment-manifest.json" file to begin the tests
with AWS Elastic Beanstalk. The idea is to create a very simple way to
deploy this web service.